Parameterize the map rendering process
Create an ``openelex`` namespace and expose a function
``openelex.renderMap()`` that kicks off the map rendering process.

This better encapsulates some of the initialization code and allows
for specifying parameters when invoking the map rendering. This
lets us specify things like the container element or the status
data file.

This also adds lodash as a dependency.  Right now it's loaded from
a CDN, like other external libraries.  I added a recipe to the
README that provides a good way to develop and test the page
locally while still using the protocol-independent CDN URLs.