Enable filtering Elections collection by office or date range
Remove the collection method that filtered by office and add one
that filters by both office and date range.

Don't pass a collection to OfficeFilterView. Instead, trigger an
event when an office is selected and let the table view handle it
and initiate the filtering.

Add some Jasmine-based tests for the new behaviors.

Addresses https://github.com/openelections/core/issues/165