Hide overlapping years
When there are very few elections for a particular year, the
next year's label will overlap this.

Try to work around this by giving the year labels a background
color so the text doesn't overlap and give the label elements a
title so we can see the full year on mouseover.

Addresses
https://github.com/openelections/openelections.github.io/issues/8