Reposition the elections table filters
Reposition the elections table filters so they always stack below
each other instead of shifting in position depending on screen
with and length of the table title.

Slightly decrease the distance between the filter widgets and
increase the distance between the widgets and the table heading
below.

Move the "Filter by" label to the top filter widget ("Race Type")

Addresses
https://github.com/openelections/openelections.github.io/issues/16