Get rid of jenky cache-busting strategy
Due to the limitations of Jekyll + GitHub pages, I had been manually
changing the filename of the state status JSON.  This is cumbersome
and makes the metadata updating process harder to document.  If it
turns out that we need it, we should come up with a more robust
cache-busting solution.  There's an issue open for this at
https://github.com/openelections/openelections.github.io/issues/42

For now, remove the datestamp from the state_status.json filename.