Allow toggling of race count legend
Allow a visitor to force-hide the "# of Races" legend on the
results page by adding the 'hidecountlegend' query parameter,
for example, '/results/#ut?hidecountlegend'.

This is to make it easier for @SaraSchnadt to see if this is
needed when showing to users at ONA.  We should eventually
decide whether to have it or not.