Adjust position of slider help text
Another adjustment, after re-reading @SaraSchnadt's request:

> can you move the text up a little, like so it's 30% down within the
> vertical space between the dates and the table title?

The previous commit moved the text so it was down 30% of the vertical
space between the slider bar and the table title.

The position is now 30% (30px) of the distance between the bottom of
the date labels and the table title (100px).

Addresses https://github.com/openelections/openelections.github.io/issues/47