Fix the map
The map wasn't rendering on initial page load.  I realized that
the same variable name `path` was being used for both the d3
path generator and the d3 selector for `path` SVG elements. It
seems like it should have worked, but it's sloppy coding to use
the same variable name anyway.