Fix URLs in home page results map
Clicking on a state currently directs the user to a URL that times out.

`/results/` works as expected, but `/results` (without the trailing
slash) redirects to <http://www.openelections.net:8001/results/>, which
simply times out.

Since the site's nav bar points to `/results/`, this just updates the
map to do the same.